move addMember to lib
