+++ server/api/member/member.controller.js
@@ -1,17 +1,7 @@
 const Member = require('./member')
 // const Person = require('../person/person')
 const Organisation = require('../organisation/organisation')
-// const { config } = require('../../../config/config')
-// const { MemberStatus } = require('./member.constants')
-// const { emailPerson } = require('../person/person.email')
-
-/* get a single member record with org and person populated out */
-const getMemberbyId = id => {
-  return Member.findOne({ _id: id })
-    .populate({ path: 'person', select: 'nickname name imgUrl email' })
-    .populate({ path: 'organisation', select: 'name imgUrl category' })
-    .exec()
-}
+const { getMemberbyId } = require('./member.lib')
 
 /**
   api/members -> list all members
@@ -68,21 +58,6 @@ const updateMember = async (req, res) => {
   }
 }
 
-// creates a new member or updates status of existing member
-const addMember = async (member) => {
-  const found = await Member.findOneAndUpdate(
-    { // check for a match
-      person: member.person,
-      organisation: member.organisation
-    },
-    member, // create or upsert
-    { new: true, upsert: true }
-  )
-  // get populated out member record
-  const got = await getMemberbyId(found._id)
-  return got
-}
-
 const createMember = async (req, res) => {
   const newMember = new Member(req.body)
   newMember.save(async (err, saved) => {
@@ -131,6 +106,5 @@ const createMember = async (req, res) => {
 module.exports = {
   listMembers,
   updateMember,
-  createMember,
-  addMember
+  createMember
 }
